refactor(ecs): move tag.hpp to system dir
